@@ -1,5 +1,5 @@
 {
-  "name": "tr",
+  "name": "@ts-2/source",
   "version": "0.0.0",
   "license": "MIT",
   "scripts": {
@@ -9,32 +9,41 @@
   },
   "private": true,
   "dependencies": {
-    "body-parser": "^1.19.0",
-    "express": "4.17.1",
-    "mongoose": "^5.13.8",
-    "tslib": "^2.0.0",
-    "uuid": "^8.3.2"
+    "axios": "^1.6.0",
+    "body-parser": "1.20.3",
+    "express": "~4.18.1",
+    "mongoose": "8.9.3",
+    "uuid": "11.0.4"
   },
   "devDependencies": {
-    "@nrwl/cli": "12.10.0",
-    "@nrwl/eslint-plugin-nx": "12.10.0",
-    "@nrwl/express": "12.10.0",
-    "@nrwl/jest": "12.10.0",
-    "@nrwl/linter": "12.10.0",
-    "@nrwl/node": "12.10.0",
-    "@nrwl/tao": "12.10.0",
-    "@nrwl/workspace": "12.10.0",
-    "@types/express": "4.17.0",
-    "@types/jest": "27.0.2",
-    "@types/node": "14.14.33",
-    "@typescript-eslint/eslint-plugin": "~4.31.1",
-    "@typescript-eslint/parser": "~4.31.1",
-    "eslint": "7.22.0",
-    "eslint-config-prettier": "8.1.0",
-    "jest": "27.2.3",
-    "prettier": "^2.3.1",
-    "supertest": "^6.1.6",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.3.5"
+    "@eslint/js": "^9.8.0",
+    "@nx/esbuild": "20.3.1",
+    "@nx/eslint": "20.3.1",
+    "@nx/eslint-plugin": "20.3.1",
+    "@nx/jest": "20.3.1",
+    "@nx/js": "20.3.1",
+    "@nx/node": "20.3.1",
+    "@nx/workspace": "20.3.1",
+    "@swc-node/register": "~1.9.1",
+    "@swc/core": "~1.5.7",
+    "@swc/helpers": "~0.5.11",
+    "@types/body-parser": "1.19.5",
+    "@types/express": "~4.17.13",
+    "@types/jest": "^29.5.12",
+    "@types/node": "~18.16.9",
+    "@types/supertest": "6.0.2",
+    "esbuild": "^0.19.2",
+    "eslint": "^9.8.0",
+    "eslint-config-prettier": "^9.0.0",
+    "jest": "^29.7.0",
+    "jest-environment-node": "^29.7.0",
+    "nx": "20.3.1",
+    "prettier": "^2.6.2",
+    "supertest": "7.0.0",
+    "ts-jest": "^29.1.0",
+    "ts-node": "10.9.1",
+    "tslib": "^2.3.0",
+    "typescript": "~5.6.2",
+    "typescript-eslint": "^8.13.0"
   }
 }
