@@ -1,10 +1,10 @@
-import {app} from '../app'
-import * as request from "supertest"
+import { app } from '../app';
+import request from 'supertest';
 
 describe('Test default route', () => {
   it('should be successful', async () => {
-    const {body, status} = await request(app).get("/api");
+    const { body, status } = await request(app).get('/api');
     expect(status).toBe(200);
-    expect(body.message).toEqual("Welcome to tournament!");
-  })
-})
+    expect(body.message).toEqual('Welcome to tournament!');
+  });
+});
